refactor(workspace): rename content/ to timeline/, rename files for clarity
Rename content/ → timeline/ to match the core abstraction. Rename files:
entry-types.ts → entries.ts, conversions.ts → richtext.ts,
sheet-csv.ts → sheet.ts, conversions.test.ts → richtext.test.ts.
Update all internal and external import paths.
